Add Gulpfile with Browserify
