+++ package.json
@@ -13,9 +13,19 @@
   },
   "homepage": "https://github.com/canalplus/r7extlib",
   "devDependencies": {
-    "mocha": "~1.17.1",
     "chai": "~1.8.1",
+    "gulp": "^3.8.10",
+    "gulp-browserify": "^0.5.0",
+    "gulp-concat": "^2.4.2",
+    "gulp-jshint": "^1.9.0",
+    "gulp-mocha": "^2.0.0",
+    "gulp-mocha-phantomjs": "^0.5.1",
+    "gulp-rimraf": "^0.1.1",
+    "lodash": "^2.4.1",
+    "mocha": "~1.17.1",
+    "q": "^1.1.2",
+    "sinon": "~1.7.3",
     "sinon-chai": "~2.4.0",
-    "sinon": "~1.7.3"
+    "yargs": "^1.3.3"
   }
 }
