+++ specs/index.html
@@ -1,30 +1,27 @@
+<!doctype html>
 <html>
-<head>
-  <meta charset="utf-8">
-  <title>r7extlib Tests</title>
-  <link rel="stylesheet" href="../bower_components/mocha/mocha.css" />
-</head>
-<body>
-  <div id="mocha"></div>
-
-  <script src="../node_modules/mocha/mocha.js"></script>
-  <script>mocha.setup('bdd');</script>
-  <script>mocha.checkLeaks();</script>
-
-  <!-- chai for assertions -->
-  <script src="../node_modules/chai/chai.js"></script>
-  <script>expect = chai.expect;</script>
-
-  <!-- sinon for mocks and stubs -->
-  <script src="../node_modules/sinon-chai/lib/sinon-chai.js"></script>
-  <script src="../node_modules/sinon/pkg/sinon.js"></script>
-  <script src="../lib/r7extlib.js"></script>
-  <script src="./r7extlib_spec.js"></script>
-
-  <script>
-    mocha.checkLeaks();
-    mocha.globals(['R7']);
-    mocha.run();
-  </script>
-</body>
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Mocha Test Runner</title>
+    <link rel="stylesheet" href="../node_modules/mocha/mocha.css">
+  </head>
+  <body>
+    <div id="mocha"></div>
+    <script src="../node_modules/mocha/mocha.js"></script>
+    <script src="../node_modules/chai/chai.js"></script>
+    <script src="../node_modules/sinon-chai/lib/sinon-chai.js"></script>
+    <script src="../node_modules/sinon/pkg/sinon.js"></script>
+    <script>mocha.setup('bdd')</script>
+    <script src="../dist/r7extlib.js"></script>
+    <script src="r7extlib_spec.js"></script>
+    <script>
+      if (window.mochaPhantomJS) {
+        mochaPhantomJS.run();
+      } else {
+        mocha.run();
+      }
+    </script>
+  </body>
 </html>
